pyproject.toml
[project]
name = "gerenciador_de_vendas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard,standart]>=0.115.6",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
]
